+++ core/src/main/java/io/snabble/sdk/Project.java
@@ -21,6 +21,7 @@ import java.util.Set;
 import java.util.concurrent.CopyOnWriteArrayList;
 
 import io.snabble.sdk.encodedcodes.EncodedCodesOptions;
+import io.snabble.sdk.utils.IntRange;
 import io.snabble.sdk.utils.JsonUtils;
 import io.snabble.sdk.utils.SimpleActivityLifecycleCallbacks;
 import okhttp3.OkHttpClient;
@@ -55,6 +56,7 @@ public class Project {
     private boolean verifyInternalEanChecksum;
     private BarcodeFormat[] supportedBarcodeFormats;
     private Shop checkedInShop;
+    private Map<BarcodeFormat, IntRange> barcodeFormatRanges;
 
     private Map<String, String> urls;
 
@@ -144,11 +146,27 @@ public class Project {
 
         useGermanPrintPrefix = JsonUtils.getBooleanOpt(jsonObject, "useGermanPrintPrefix", false);
 
+        String[] scanFormats = JsonUtils.getStringArrayOpt(jsonObject, "scanFormats", null);
         List<BarcodeFormat> formats = new ArrayList<>();
-        formats.add(BarcodeFormat.EAN_8);
-        formats.add(BarcodeFormat.EAN_13);
-        formats.add(BarcodeFormat.CODE_128);
-        //formats.add(BarcodeFormat.ITF);
+
+        if(scanFormats != null) {
+            for (String scanFormat : scanFormats) {
+                BarcodeFormat format = BarcodeFormat.parse(scanFormat);
+                if(format != null) {
+                    formats.add(format);
+                }
+            }
+        } else {
+            formats.add(BarcodeFormat.EAN_8);
+            formats.add(BarcodeFormat.EAN_13);
+            formats.add(BarcodeFormat.CODE_128);
+        }
+
+        barcodeFormatRanges = new HashMap<>();
+        // TODO parse from metadata
+        if (id.contains("ikea")) {
+            barcodeFormatRanges.put(BarcodeFormat.ITF_14, new IntRange(0, 8));
+        }
 
         supportedBarcodeFormats = formats.toArray(new BarcodeFormat[formats.size()]);
 
@@ -322,6 +340,14 @@ public class Project {
 
     public Checkout getCheckout() {
         return checkout;
+    }
+
+    public IntRange getRangeForBarcodeFormat(BarcodeFormat barcodeFormat) {
+        if (barcodeFormatRanges != null) {
+            return barcodeFormatRanges.get(barcodeFormat);
+        }
+
+        return null;
     }
 
     Events getEvents() {