+++ sample/src/main/java/io/snabble/testapp/HomeFragment.java
@@ -1,9 +1,12 @@
 package io.snabble.testapp;
 
+import android.content.Intent;
 import android.os.Bundle;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+import io.snabble.sdk.ui.receipts.ReceiptListView;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -31,6 +34,14 @@ public class HomeFragment extends Fragment {
             @Override
             public void onClick(View v) {
                 ((BaseActivity)getActivity()).showShoppingCart();
+            }
+        });
+
+        v.findViewById(R.id.receipt_list).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(requireActivity(), ReceiptListActivity.class);
+                startActivity(intent);
             }
         });