+++ ui/src/main/java/io/snabble/sdk/ui/scanner/ZXingHelper.java
@@ -11,7 +11,9 @@ class ZXingHelper {
                 return com.google.zxing.BarcodeFormat.EAN_13;
             case CODE_128:
                 return com.google.zxing.BarcodeFormat.CODE_128;
-            case ITF:
+            case CODE_39:
+                return com.google.zxing.BarcodeFormat.CODE_39;
+            case ITF_14:
                 return com.google.zxing.BarcodeFormat.ITF;
             case DATA_MATRIX:
                 return com.google.zxing.BarcodeFormat.DATA_MATRIX;
@@ -30,8 +32,10 @@ class ZXingHelper {
                 return BarcodeFormat.EAN_13;
             case CODE_128:
                 return BarcodeFormat.CODE_128;
+            case CODE_39:
+                return BarcodeFormat.CODE_39;
             case ITF:
-                return BarcodeFormat.ITF;
+                return BarcodeFormat.ITF_14;
             case DATA_MATRIX:
                 return BarcodeFormat.DATA_MATRIX;
             case QR_CODE: