+++ utils/src/main/java/io/snabble/sdk/utils/IntRange.java
@@ -1,0 +1,16 @@
+package io.snabble.sdk.utils;
+
+public class IntRange {
+    public final int min;
+    public final int max;
+
+    public IntRange() {
+        this.min = 0;
+        this.max = 0;
+    }
+
+    public IntRange(int min, int max) {
+        this.min = min;
+        this.max = max;
+    }
+}